v0.0.5
In v0.0.5, modules can now have multi-character names! 

Simply add to the handle_extern function, an if statement, with if i is equal to followed by a number. 

That number is the module's name, written like this. 

If you have the string 'about', 

It becomes:
([ASCII VALUE OF a] * (pow(256,3))) + ([ASCII VALUE OF b] * (pow(256,2))) + ([ASCII VALUE OF o] * (pow(256,1))) + ([ASCII VALUE OF u])

And you simply use that formula for whatever string you want.