ADD CALCULATOR!
Type add for addition, sub for subtraction, and mlt for multiplication. 
Syntax:
electron# add

electron# 02 22
42

This may look incorrect, but this is because it is displayed backwards, as I have not gotten reversing a string working yet. 
And the 02, is because they need to be the same length. 

Pad the left number by placing zero's to the left, until it is the same size. 
If the left is larger, place zero's to the RIGHT's left, until it is the same size.